pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigdiscovpy"
version = "0.1.0"
description = "GPU-accelerated spatial signature discovery for spatial transcriptomics"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Seongyong Park", email = "seongyong.park@example.com"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["spatial-transcriptomics", "moran", "spatial-statistics", "gpu", "cupy"]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "anndata>=0.8.0",
    "h5py>=3.0.0",
    "tqdm>=4.60.0",
    "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x>=12.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
]
dev = [
    "sigdiscovpy[test]",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "build>=1.0.0",
]
all = [
    "sigdiscovpy[gpu,test,dev]",
]

[project.urls]
Homepage = "https://github.com/psychemistz/sigdiscovPy"
Repository = "https://github.com/psychemistz/sigdiscovPy"
Issues = "https://github.com/psychemistz/sigdiscovPy/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sigdiscovpy*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = [
    "E501",  # Line too long (handled by black)
    "E741",  # Ambiguous variable name (I, l, O) - common in math notation
    "B904",  # raise ... from err - less critical
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
